backend/pyproject.toml
[tool.poetry]
name = "northstar-backend"
version = "0.1.0"
description = ""
authors = ["Reid Wyde <reidwyde@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
google-api-python-client = "^2.161.0"
google-auth-httplib2 = "^0.2.0"
google-auth-oauthlib = "^1.2.1"
pyjwt = "^2.8.0"
cryptography = "^41.0.0"
requests = "^2.31.0"
boto3 = "^1.34.0"


[tool.poetry.scripts]
trigger = "tests.trigger_lambda:trigger_lambda"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
